reorganized benchmark and test headers
